Split `console_integration.rs` from `console_repl.rs`
